extract sankey defs
