+++ frontend/scripts/react-components/shared/shrinking-spinner/shrinking-spinner.scss
@@ -1,13 +1,12 @@
 @import 'styles/settings';
 
-$width-small: 18px;
-$width-regular: 30px;
+$width-small: 30px;
 $width-large: 60px;
 
 .c-shrinking-spinner {
   position: relative;
   margin: 0 auto;
-  width: $width-regular;
+  width: $width-small;
 
   &:before {
     content: '';
@@ -17,10 +16,6 @@ $width-large: 60px;
 
   &.-large {
     width: $width-large;
-  }
-
-  &.-small {
-    width: $width-small;
   }
 
   &.-dark {
