+++ frontend/scripts/react-components/tool/columns-selector-group/columns-selector-group.component.jsx
@@ -1,15 +1,23 @@
-import React from 'react';
+import React, { useState, useLayoutEffect } from 'react';
 import PropTypes from 'prop-types';
 import ColumnSelector from 'react-components/tool/column-selector/column-selector.container';
 import Heading from 'react-components/shared/heading';
 import cx from 'classnames';
 
 import 'react-components/tool/columns-selector-group/columns-selector-group.scss';
-import ShrinkingSpinner from 'react-components/shared/shrinking-spinner/shrinking-spinner.component';
 
 function ColumnsSelectorGroup(props) {
   const { sankeySize, columns } = props;
   const loading = !columns || columns.length === 0;
+  const [dots, setDots] = useState(1);
+  useLayoutEffect(() => {
+    const interval = setInterval(() => setDots(d => (d + 1) % 4), 300);
+    return () => {
+      if (interval) {
+        clearInterval(interval);
+      }
+    };
+  });
 
   const styles = { width: `${sankeySize[0] + 8}px` };
 
@@ -19,12 +27,9 @@ function ColumnsSelectorGroup(props) {
       className={cx('c-columns-selector-group', 'is-absolute', { '-loading': loading })}
     >
       {loading && (
-        <>
-          <Heading variant="mono" size="sm" weight="bold">
-            Loading...
-          </Heading>
-          <ShrinkingSpinner className="-dark columns-spinner -small" />
-        </>
+        <Heading variant="mono" size="md" weight="bold">
+          Loading{Array.from({ length: dots }).map(() => '.')}
+        </Heading>
       )}
 
       {!loading && Array.from({ length: 4 }).map((_, i) => <ColumnSelector group={i} />)}
