+++ frontend/styles/components/tool/sankey.scss
@@ -92,9 +92,4 @@
     stroke-width: 1px;
     fill: transparent;
   }
-
-  .sankey-link-placeholder {
-    fill: $charcoal-dark;
-    opacity: 0.5;
-  }
 }
